requirements/base.txt
-r ./dev.txt

boto3~=1.9.164
flask-cors~=3.0.7
flask-login~=0.4.1
flask-marshmallow~=0.10.1
flask-migrate~=2.5.0
flask-rest-api~=0.16.1
flask-sqlalchemy~=2.4.0
flask~=1.1.0
marshmallow-enum~=1.4.1
marshmallow-jsonschema~=0.6.0
marshmallow-sqlalchemy~=0.17.0
phonenumbers~=8.10.0
pymysql~=0.9.3
python-dotenv~=0.10.3
qcloudsms-py~=0.1.3
requests~=2.22.0
sqlalchemy-utils~=0.34.0
tencentcloud-sdk-python~=3.0.71
webargs~=5.4.0
